back/requirements.txt
pandas 
gspread
python-dotenv
flask
flask_cors
oauth2client